fix: Correctness fixes in covar1, fdjac1, lmpar, qrfac, dogleg; bump to 1.3.13
- covar1: guard the residual-variance division fsumsq/(m-k) against m <= rank
  (was a divide-by-zero / 0-0 producing Inf/NaN over the whole covariance
  matrix, e.g. on square full-rank problems). covar1 is a C-only routine with
  no f2c counterpart.
- fdjac1: the banded finite-difference branch now calls the user function with
  iflag=2, matching the dense branch, fdjac2 and the FORTRAN version (where
  hybrd sets iflag=2 and fdjac1 forwards it). Update the stale iflag docs.
- lmpar (USE_BLAS): the Newton correction now uses all n components; S is full
  rank because par>0, so truncating to nsing gave a wrong step for
  rank-deficient Jacobians.
- qrfac (USE_LAPACK): fix the jpvt memset size (sizeof(*jpvt), not sizeof(int))
  and document the malloc/info error-handling limitation.
- Make the work-array size checks in hybrd, hybrj, hybrd1, hybrj1 and lmdif1
  overflow-safe (size_t), in both the pure-C and f2c versions.
- Guard a harmless transient infinity in dogleg for rank-deficient Jacobians,
  in both the pure-C and f2c versions (behaviour-preserving).
- Document the double-evaluation caveat of the min/max/abs macros.
- Bump version to 1.3.13 and add the changelog entry.

+++ src/f2c/hybrd1_.c
@@ -119,7 +119,7 @@ void __minpack_func__(hybrd1)(__minpack_decl_fcn_nn__  const int *n, real *x, re
 
 /*     check the input parameters for errors. */
 
-    if (*n <= 0 || *tol < 0. || *lwa < *n * (*n * 3 + 13) / 2) {
+    if (*n <= 0 || *tol < 0. || (size_t)*lwa < (size_t)*n * (*n * 3 + 13) / 2) {
 	/* goto L20; */
         return;
     }
