+++ src/hybrd1.c
@@ -115,7 +115,7 @@ int __cminpack_func__(hybrd1)(__cminpack_decl_fcn_nn__ void *p, int n, real *x, 
 
 /*     check the input parameters for errors. */
 
-    if (n <= 0 || tol < 0. || lwa < n * (n * 3 + 13) / 2) {
+    if (n <= 0 || tol < 0. || (size_t)lwa < (size_t)n * (n * 3 + 13) / 2) {
         return 0;
     }
